refactor: Marketplace `AppsPage` structure
