+++ apps/meteor/client/views/marketplace/AppsPage/AppsPageContentBody.tsx
@@ -1,8 +1,9 @@
 import type { App } from '@rocket.chat/core-typings';
-import { Box, Pagination } from '@rocket.chat/fuselage';
+import { Box } from '@rocket.chat/fuselage';
 import { useUniqueId } from '@rocket.chat/fuselage-hooks';
 import type { PaginatedResult } from '@rocket.chat/rest-typings';
-import React, { useRef } from 'react';
+import type { RefObject } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 
 import AppsList from '../AppsList';
@@ -17,58 +18,18 @@ type AppsPageContentBodyProps = {
 		allApps: App[];
 		totalAppsLength: number;
 	}>;
-	itemsPerPage: 25 | 50 | 100;
-	current: number;
-	onSetItemsPerPage: React.Dispatch<React.SetStateAction<25 | 50 | 100>>;
-	onSetCurrent: React.Dispatch<React.SetStateAction<number>>;
-	paginationProps: {
-		itemsPerPageLabel: () => string;
-		showingResultsLabel: (context: { count: number; current: number; itemsPerPage: 25 | 50 | 100 }) => string;
-	};
-	noErrorsOcurred: boolean;
+	scrollableRef: RefObject<HTMLDivElement>;
 };
 
-const AppsPageContentBody = ({
-	isMarketplace,
-	isFiltered,
-	appsResult,
-	itemsPerPage,
-	current,
-	onSetItemsPerPage,
-	onSetCurrent,
-	paginationProps,
-	noErrorsOcurred,
-}: AppsPageContentBodyProps) => {
+const AppsPageContentBody = ({ isMarketplace, isFiltered, appsResult, scrollableRef }: AppsPageContentBodyProps) => {
 	const { t } = useTranslation();
-	const scrollableRef = useRef<HTMLDivElement>(null);
 	const appsListId = useUniqueId();
 
 	return (
-		<>
-			<Box display='flex' flexDirection='column' overflow='hidden' height='100%' pi={24}>
-				{noErrorsOcurred && (
-					<Box overflowY='scroll' height='100%' ref={scrollableRef}>
-						{isMarketplace && !isFiltered && <FeaturedAppsSections appsListId={appsListId} appsResult={appsResult?.allApps || []} />}
-						<AppsList appsListId={appsListId} apps={appsResult?.items || []} title={isMarketplace ? t('All_Apps') : undefined} />
-					</Box>
-				)}
-			</Box>
-			{Boolean(appsResult?.count) && (
-				<Pagination
-					divider
-					current={current}
-					itemsPerPage={itemsPerPage}
-					count={appsResult?.total || 0}
-					onSetItemsPerPage={onSetItemsPerPage}
-					onSetCurrent={(value) => {
-						onSetCurrent(value);
-						scrollableRef.current?.scrollTo(0, 0);
-					}}
-					bg='light'
-					{...paginationProps}
-				/>
-			)}
-		</>
+		<Box overflowY='scroll' height='100%' ref={scrollableRef}>
+			{isMarketplace && !isFiltered && <FeaturedAppsSections appsListId={appsListId} appsResult={appsResult?.allApps || []} />}
+			<AppsList appsListId={appsListId} apps={appsResult?.items || []} title={isMarketplace ? t('All_Apps') : undefined} />
+		</Box>
 	);
 };
